refactor(consensus): extract useConsensusReconciliation from ExtractionFullScreen (file-size gate)
